设置存款打印位数

+++ include/trade_conclude.h
@@ -17,8 +17,8 @@
 typedef struct
 {
     int customers_num;  ///< 接待客户总数
-    int flowing_water;  ///< 总流水
-    int total_deposits; ///< 总存款量
+    double flowing_water;  ///< 总流水
+    double total_deposits; ///< 总存款量
 } trade_conclude;
 
 /**
